SparseArray 1.1.8: Add aperm() method for SVT_SparseArray objects
The aperm() method for SVT_SparseArray objects is currently restricted to
permutations that transpose the **leftmost dimensions** (a.k.a. innermost
or fastest moving dimensions in the memory layout). In other words, it only
supports a 'perm' argument that is identical to 'c(r:1, (r+1):ndim)',
where 'r' is the "perm rank". See src/SparseArray_aperm.c for the details.