Bump SVT_VERSION from 0 to 1
The layout of the Sparse Vector Tree (SVT) used in SVT_SparseArray objects
is now versioned, and internal global variable SVT_VERSION keeps track of
that.
In SVT_VERSION 1, the order of the nzvals/nzoffs components in the SVT
leaves is switched w.r.t. SVT_VERSION 0: new order is nzvals/nzoffs;
previous order was nzoffs/nzvals.
Version 0 was used in SparseArray < 1.5.0. This means that SVT_SparseArray
objects created and serialized with SparseArray < 1.5.0 are not compatible
with SparseArray >= 1.5.0 for now. An updateObject() method might be added
later.