Add a couple of nzwhich()/nzvals() methods
Add a nzwhich() method for RsparseMatrix objects and nzvals() methods
for ngCMatrix and ngRMatrix objects.

Also brings back coercion from dgCMatrix to ngCMatrix. This is another
critically endangered coercion method originally defined in the Matrix
package but that the lazy Matrix maintainers have decided to eradicate
from the surface of Earth in an attempt to make the life of their users
no so easy.