Implement `[<-`(<SVT_SparseArray>, ..., value=<SVT_SparseArray>)
Subassignment of an SVT_SparseArray object with a right value that is
also an SVT_SparseArray object is finally supported. For example this
allows things like

    x[3:6, , 1] <- y
    x[ , 11:15, 1] <- x[ , 1:5, 3]

(where 'x' and 'y' are both SVT_SparseArray objects) to operate
natively on the sparse representation and therefore to be very fast
and very memory-efficient.